requirements.txt
pyTelegramBotAPI==4.15.4
requests==2.31.0
feedparser==6.0.11
supabase==2.4.5
python-dotenv==1.0.1
pytz==2024.1
httpx==0.27.0
